test: fix #815 guardrail and coverage gates
